Canvas2D Text: Use a managed child node to render text
- Canvas Text Textures are now managed properly by the texture manager
- Also add willReadFrequently option to text drawing canvas element context
