+++ src/core/CoreTextNode.ts
@@ -390,6 +390,15 @@ export class CoreTextNode
   override renderQuads(renderer: CoreRenderer) {
     assertTruthy(this.globalTransform);
 
+    // If the text renderer does not support rendering quads, fallback to the
+    // default renderQuads method
+    if (!this.textRenderer.renderQuads) {
+      super.renderQuads(renderer);
+      return;
+    }
+
+    // If the text renderer does support rendering quads, use it...
+
     // Prevent quad rendering if parent has a render texture
     // and this node is not the render texture
     if (this.parentHasRenderTexture) {
@@ -444,7 +453,7 @@ export class CoreTextNode
       this._textRendererOverride,
     );
 
-    const textRendererState = resolvedTextRenderer.createState(props);
+    const textRendererState = resolvedTextRenderer.createState(props, this);
 
     textRendererState.emitter.on('loaded', this.onTextLoaded);
     textRendererState.emitter.on('failed', this.onTextFailed);
