+++ src/core/textures/Texture.ts
@@ -221,14 +221,6 @@ export abstract class Texture extends EventEmitter {
     return ctxTexture;
   }
 
-
-  /**
-   * Free the texture
-   */
-  free(): void {
-    this.setState('freed');
-  }
-
   /**
    * Set the state of the texture
    *
